requirements.txt
asgiref==3.7.2
Django==4.2.7
django-autoslug==1.9.9
django-colorfield==0.11.0
pillow==10.2.0
sqlparse==0.4.4
tzdata==2023.3
